% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_linebreaks_attributes}
\alias{add_linebreaks_attributes}
\title{Adds line breaks to forge axes labels}
\usage{
add_linebreaks_attributes(labels)
}
\arguments{
\item{labels}{a vector of strings}
}
\value{
a vector of strings
}
\description{
In case labels refer to several attributes, such as "a=1,b=2", replaces the commas by
carriage returns
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
