% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{as.data.frame.dpp_degree_cpt}
\alias{as.data.frame.dpp_degree_cpt}
\title{Coerce a distribution of degrees into a data frame}
\usage{
\method{as.data.frame}{dpp_degree_cpt}(x, ...)
}
\arguments{
\item{x}{the distribution of degrees created by \code{\link{create_degree_probabilities_table}}}

\item{...}{further parameters are ignored quietly}
}
\description{
Extracts the data frame of the distribution of degrees.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
