% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{as.data.frame.dpp_matching_probas}
\alias{as.data.frame.dpp_matching_probas}
\title{Coerce pairing probabilities into a data frame}
\usage{
\method{as.data.frame}{dpp_matching_probas}(x, ...)
}
\arguments{
\item{x}{the pairing probabilities to convert}

\item{...}{further parameters are ignored quietly}
}
\description{
Extracts the data frame of pairing probabilities prepared by function \code{\link{create_matching_probabilities_table}}
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
