% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{as.data.frame.dpp_sample}
\alias{as.data.frame.dpp_sample}
\title{Coerce the a dpp sample into a data frame}
\usage{
\method{as.data.frame}{dpp_sample}(x, ...)
}
\arguments{
\item{x}{the sample to convert}

\item{...}{further parameters are ignored quietly}
}
\description{
Extracts the data frame of a sample prepare by function \code{\link{create_sample}}
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
