% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as.igraph.dpp_population}
\alias{as.igraph.dpp_population}
\title{Exports a population as an igraph graph.}
\usage{
\method{as.igraph}{dpp_population}(pop, with.attributes = FALSE, ...)
}
\arguments{
\item{pop}{the population to convert}

\item{with.attributes}{if TRUE, the attributes of entities are copied as vertex attributes}

\item{...}{ignored}
}
\value{
an igraph graph
}
\description{
The resulting graph is directed. 
Depending to parameters its vertices might be attributed.
}
\examples{

library(igraph)
# generate a population based on sample case 1
data(dwellings_households)
prepared <- matching.prepare(
                    dwellings_households$sample.A, dwellings_households$sample.B, 
                    dwellings_households$pdi, dwellings_households$pdj, 
                    dwellings_households$pij)
solved <- matching.solve(prepared, 
                    nA=500,nB=400, 
                    nu.A=0, phi.A=0, delta.A=1, 
                    gamma=1, 
                    delta.B=0, phi.B=0, nu.B=0)
sampled <- matching.generate(solved, dwellings_households$sample.A, dwellings_households$sample.B)

# convert the population as a population
g <- as.igraph(sampled$pop, with.attributes=TRUE)

# can export the graph using one of the igraph formats
write.graph(g,file="mygraph.graphml", format="graphml")

# can visualize the population as a graph
#lay <- layout_nicely(g)
#tkplot(g,layout=lay)

# test structural properties
is.connected(g)
average.path.length(g)

# check attributes 
# ... list all the attributes defined for verticies 
vertex_attr_names(g)
# ... view all attributes
vertex_attr(g)


}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
