% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as.igraph.dpp_result}
\alias{as.igraph.dpp_result}
\title{Exports a generation result as an igraph graph.}
\usage{
\method{as.igraph}{dpp_result}(generated, with.attributes = FALSE, ...)
}
\arguments{
\item{generated}{the generation result to convert to igraph}

\item{with.attributes}{if TRUE, the attributes of entities are copied as vertex attributes}

\item{...}{ignored}
}
\description{
Exports a generation result as an igraph graph.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
