% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as.latex.relaxation.constrainsts}
\alias{as.latex.relaxation.constrainsts}
\title{Writes relaxation parameters as LaTeX}
\usage{
as.latex.relaxation.constrainsts(sp)
}
\arguments{
\item{sp}{a solved case produced by \code{\link{matching.solve}}}
}
\value{
a string containing a LaTeX string.
}
\description{
Create LaTeX code that depicts the relaxation parameters.
}
\examples{
data(dwellings_households)
prepared <- matching.prepare(
                     dwellings_households$sample.A, dwellings_households$sample.B, 
                     dwellings_households$pdi, dwellings_households$pdj, 
                     dwellings_households$pij)
solved <- matching.solve(
                     prepared, nA=50000, nB=40000, nu.A=1, phi.A=0, 
                     delta.A=0, gamma=0, delta.B=0, phi.B=0, nu.B=1, verbose=TRUE)
# create the string
rel <- as.latex.relaxation.constrainsts(solved)
print(rel)
# write it into a file
# write(rel, , file="relaxation_parameters.tex")

}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{latex}
