% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as.latex.table.discrete}
\alias{as.latex.table.discrete}
\title{Creates a LaTeX view with the discrete values}
\usage{
as.latex.table.discrete(sp, maxcol = NULL)
}
\arguments{
\item{sp}{a solved case produced by \code{\link{matching.solve}}}

\item{maxcol}{an optional integer containing the highest count of columns to display (useful for large tables)}
}
\value{
a string containing a LaTeX table
}
\description{
Create a table view of the solved problem, with integer values. 
Ready to be written into a tex file.
}
\examples{
data(dwellings_households)
prepared <- matching.prepare(
                     dwellings_households$sample.A, dwellings_households$sample.B, 
                     dwellings_households$pdi, dwellings_households$pdj, 
                     dwellings_households$pij)
solved <- matching.solve(
                     prepared, nA=50000, nB=40000, nu.A=1, phi.A=0, 
                     delta.A=0, gamma=0, delta.B=0, phi.B=0, nu.B=1, verbose=TRUE)
# create the string
dv <- as.latex.table.discrete(solved)
print(dv)
# write it into a file
# write(dv, , file="case1_discrete.tex")

}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{latex}
