% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as.merged.dataframe}
\alias{as.merged.dataframe}
\title{Exports a synthetic population as one unique dataframe}
\usage{
as.merged.dataframe(generated)
}
\arguments{
\item{generated}{the generation result produced by \code{\link{matching.generate}}}
}
\description{
Merges the individual dataframes of population A, B and links
into one unique merged one. There will be partial lines when 
one entity had no link.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
