% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{assert.equal}
\alias{assert.equal}
\title{Asserts two values are equal.}
\usage{
assert.equal(v1, v2, msg, verbose = FALSE, indent = 3,
  tolerance = 1.5e-08)
}
\arguments{
\item{v1}{a scalar, vector, matrix or data frame}

\item{v2}{a scalar, vector, matrix or data frame}

\item{msg}{the message to display in case of failure}

\item{verbose}{if TRUE, assertion errors will be displayed in the console}

\item{indent}{the indentation (count of tabs) for verbose display}

\item{tolerance}{the numeric tolerance to define two numeric are equal}
}
\value{
1 in case of error else 0
}
\description{
In the context of checking the consistency of a solution,
ensures the content are similar.
}
\keyword{internal}
