% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{assert.hypothesis}
\alias{assert.hypothesis}
\title{Asserts a given hypothesis on a given solution for a given case}
\usage{
assert.hypothesis(hypothesis, sol, case, nA, nB, verbose = FALSE)
}
\arguments{
\item{hypothesis}{a list of strings representing variables, as generated by \code{\link{generate.hypothesis}}}

\item{sol}{a current solution}

\item{case}{the case containing the initial values}

\item{verbose}{FALSE by default
 @return the solution with more variables}
}
\description{
For an hypothesis such as l("hat.fi","hat.pdi"), sets the corresponding
variables in the solution such as sol$hat.fi=case$fi and sol$hat.pdi=case$pdi.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
