% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{assess.max.pd}
\alias{assess.max.pd}
\title{Computes the maximum degree which might be obtained using a given 
distribution of degrees. It corresponds to the maximal average degre
which can be obtained by still enforcing the zeros in the initial distribution.}
\usage{
assess.max.pd(orig.pd)
}
\arguments{
\item{orig.pd}{the probability distribution}
}
\value{
a vector containing the maximum possible degree
}
\description{
Computes the maximum degree which might be obtained using a given 
distribution of degrees. It corresponds to the maximal average degre
which can be obtained by still enforcing the zeros in the initial distribution.
}
\keyword{internal}
