% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{best.solution}
\alias{best.solution}
\title{Selects one best solution among a list of possible solutions}
\usage{
best.solution(solutions, nu.A, phi.A, delta.A, gamma, nu.B, phi.B, delta.B,
  verbose = FALSE)
}
\arguments{
\item{solutions}{a list of solutions such as generated inside \code{\link{resolve}}}

\item{nu.A}{control for nA: 0 means "respect nA", non-null "adapt it to solve the case"}

\item{phi.A}{control for frequencies: 0 means "respect the original frequencies as detected in the sample", non-null "adapt it to solve the case"}

\item{delta.A}{control for degree A: 0 means "respect the input parameters pdi", non-null "adapt them to solve the case"}

\item{gamma}{control for pij: 0 means "respect the matching probabilities pij", non-null "adapt them to solve the case"}

\item{nu.B}{control for nB: 0 means "respect nB", non-null "adapt it to solve the case"}

\item{phi.B}{control for frequencies: 0 means "respect the original frequencies as detected in the sample", non-null "adapt it to solve the case"}

\item{delta.B}{control for degree B: 0 means "respect the input parameters pdj", non-null "adapt them to solve the case"}

\item{verbose}{if TRUE, will display detailed information on the console}
}
\value{
a solution solved
}
\description{
Relying on the weigthed error of each solution, selects one of the best ones.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
