% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{compute_average_degree}
\alias{compute_average_degree}
\title{Computes the average degree from a distribution of degrees}
\usage{
compute_average_degree(pd)
}
\arguments{
\item{pd}{the probability distribution}
}
\value{
a vector containing the average degree
}
\description{
Computes the average degree from a distribution of degrees
}
