% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compute_common_columns}
\alias{compute_common_columns}
\title{Identifies the columns present it two datasets}
\usage{
compute_common_columns(t1, t2)
}
\arguments{
\item{t1}{a dataframe}

\item{t2}{a dataframe}
}
\value{
a list of strings
}
\description{
Identifies the columns present it two datasets
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
