% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{create_degree_probabilities_table}
\alias{create_degree_probabilities_table}
\title{Creates a table storing probabilities for degrees}
\usage{
create_degree_probabilities_table(probabilities, norm = TRUE)
}
\arguments{
\item{probabilities}{a data frame containing the probabilities}

\item{norm}{if TRUE, will normalize the table so the columns sum up to 1 (defaults to TRUE)}
}
\description{
Creates a table storing probabilities for degrees
}
\examples{
# create a table describing degrees depending to size; the bigger the size, the highest the degree
p <- data.frame(
                 'size=0'=c(0.2, 0.8, 0, 0, 0),
                 'size=1'=c(0.15, 0.8, 0.05, 0, 0),
                 'size=2'=c(0.05, 0.8, 0.1, 0.05, 0),
                 check.names=FALSE
                 )
create_degree_probabilities_table(probabilities=p)


}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
