% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{create_matching_probabilities_table}
\alias{create_matching_probabilities_table}
\title{Creates a table storing matching probabilities}
\usage{
create_matching_probabilities_table(data, norm = TRUE)
}
\arguments{
\item{data}{a data frame containing matching probabilities}

\item{norm}{if TRUE, will normalize the table so the totals sum up to one (defaults to TRUE)}
}
\value{
a matching probability table ready to be used for usage with \code{\link{matching.prepare}}
}
\description{
Creates a table storing matching probabilities
}
\examples{

dwellings_households.pij <- create_matching_probabilities_table(
               data.frame(
                   'surface=1'=c(0.2, 0.1, 0.05, 0.025),
                   'surface=2'=c(0.0375, 0.125, 0.1, 0.05),
                   'surface=3'=c(0.0125, 0.025, 0.1, 0.175),
                   row.names=c("size=1", "size=2", "size=3", "size=4"),
                   check.names=FALSE
                   )
               )
print(dwellings_households.pij)

}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
