% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{create_sample}
\alias{create_sample}
\title{Creates a sample organized for dpp manipulation.}
\usage{
create_sample(data, encoding = NULL, weight.colname = NULL)
}
\arguments{
\item{data}{a data frame containing a sample (weighted list) of entities}

\item{encoding}{a dictionary containing information about the variables of the sample}

\item{weight.colname}{the name of the column containing the weights}
}
\value{
a sample
}
\description{
The resulting object contains the sample and a dictionary of data. 
If no weight column is provided, one will be created with value 1 (uniform weights).
}
\examples{
# to read a CSV file as a sample
f <- system.file("extdata", "logements.csv", package = "gosp.dpp")
m <- read.csv(f, sep=";", dec=",", check.names=FALSE)
df <- as.data.frame(m, check.names=FALSE)
dictionary <- list('surface'=list('small'=1, 'medium'=2, 'large'=3))
create_sample(data=df, encoding=dictionary, weight.colname="weight")

# to create a sample from random data
# ... create 100 entities being either male of female
df <- data.frame(gender=sample(1:2, size=100, replace=TRUE))
# ... describe the encoding of data
dictionary <- list("gender"=list("male"=1,"female"=2))
create_sample(data=df, encoding=dictionary)
# ... the weights columns was created automatically

}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
