% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{detect.missing.chains}
\alias{detect.missing.chains}
\title{Detects the missing chains of variables}
\usage{
detect.missing.chains(sol, verbose = FALSE)
}
\arguments{
\item{sol}{the current solution (a named list)}

\item{verbose}{if TRUE, will display detailed information on the console}
}
\value{
a list of vectors (the chains) of strings
}
\description{
For a given temptative solution supposed to be consistent, 
checks the successive variables which depend on each other and 
are not yet solved. These chains of variables probably require 
using hypothesis for actual resolution.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
