% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{detect.problems}
\alias{detect.problems}
\title{Checks the consistency of a solution}
\usage{
detect.problems(sol, case, fail = TRUE, verbose = FALSE, indent = 1,
  tolerance.pd = 1.5e-08, tolerance.degree.max = 1.5e-08,
  tolerance.pij = 1e-04)
}
\arguments{
\item{sol}{the current solution (a named list)}

\item{case}{the case to be solved}

\item{fail}{if TRUE, an error will call stop()}

\item{verbose}{if TRUE, will ddetect.problemsisplay detailed information on the console}

\item{indent}{the indentation (count of tabs) for verbose display}

\item{tolerance.pd}{the tolerance for probability distributions}

\item{tolerance.degree.max}{the tolerance for min/max average degrees}

\item{tolerance.pij}{the tolerance for pij probabilities}
}
\value{
the count of problems
}
\description{
Takes a temptative solution and the initial case to resolve. 
Tests wether the equations all lead to consistent results of 
do contradict each other. Returns an error with \code{stop()}
in case of error if \code{fail=TRUE}; else just returns the 
count of problems.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
