% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwellings_households}
\alias{dwellings_households}
\title{Example dataset for DPP with simple dwellings and households}
\format{An object of class \code{list} of length 5.}
\usage{
dwellings_households
}
\description{
This is fake, forged data for testing purpose. Contains a sample of dwellings, 
and sample of households, and the data required to associate them together.
}
\keyword{datasets}
