% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{ensure.form}
\alias{ensure.form}
\title{Ensures the content of a solution is in the expect format}
\usage{
ensure.form(sol, case)
}
\arguments{
\item{sol}{the solution (might be partial)}

\item{case}{the case}
}
\value{
the same solution fixed formats
}
\description{
Ensures a solution is formatted well: notably guarantees
the lists or data frames which have to be names are named 
as expected.
}
\keyword{internal}
