% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ensure_presence_ggplot}
\alias{ensure_presence_ggplot}
\title{Stops if ggplot2 is not installed}
\usage{
ensure_presence_ggplot()
}
\description{
Stops if ggplot2 is not installed
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
