% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{errors.pdj}
\alias{errors.pdj}
\title{Measure the errors in the distributions of degrees pdj}
\usage{
errors.pdj(sp)
}
\arguments{
\item{sp}{a synthetic population, as produced by \code{\link{matching.generate}} or \code{\link{matching.solve}}.}
}
\value{
a dataframe containing the difference between the expected and actual distribution of degrees
}
\description{
Returns the difference between the actual and input distributions of degrees. 
For instance a "minus one" means we reached one less than expected.
}
\seealso{
\code{\link{errors.pdi}} for the measure of distributions of degrees pdi
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
