% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{errors.pij}
\alias{errors.pij}
\title{Measure the errors in the pairing probabilities  pij}
\usage{
errors.pij(sp)
}
\arguments{
\item{sp}{a synthetic population, as produced by \code{\link{matching.generate}} or \code{\link{matching.solve}}.}
}
\value{
a dataframe containing the difference between the expected and actual pairing probabilities
}
\description{
Returns the difference between the actual and input pairing probabilities. 
For instance a "minus one" means we reached one less than expected.
}
\seealso{
\code{\link{errors.pdi}} for the measure of distributions of degrees pdi
			\code{\link{errors.pdj}} for the measure of distributions of degrees pdj
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
