% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{expand_with_variable.dpp_degree_cpt}
\alias{expand_with_variable.dpp_degree_cpt}
\title{Expands a conditional probability table for degrees with an additional variable}
\usage{
expand_with_variable.dpp_degree_cpt(cpt, name, values)
}
\arguments{
\item{cpt}{a conditional probability table for degrees}

\item{name}{a string}

\item{values}{a list of values}
}
\value{
the CPT with additional columns minus the original columns
}
\description{
Takes the conditional probability table for degrees passed as a parameter, and 
adds to columns the combinations of existing values and the novel
variable values. 
For instance if the columns where a=1 and a=2, and you add name=b
and values=list(3,4), the columns will be a=1,b=3 a=1,b=4 a=2,b=3
a2,b=4
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
