% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{expand_with_variable.dpp_matching_probas.A}
\alias{expand_with_variable.dpp_matching_probas.A}
\title{Expands the columns of matching probabilities with additional variables}
\usage{
expand_with_variable.dpp_matching_probas.A(pt, name, values)
}
\arguments{
\item{pt}{a probability table for degrees}

\item{name}{a string}

\item{values}{a list of values}
}
\value{
the CPT with additional columns minus the original columns
}
\description{
Takes the probability table for matching passed as a parameter, 
and extends its columns with a novel variable
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
