% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{expand_with_variable.dpp_matching_probas.B}
\alias{expand_with_variable.dpp_matching_probas.B}
\title{Expands the rows of matching probabilitis with additional variables}
\usage{
expand_with_variable.dpp_matching_probas.B(pt, name, values)
}
\arguments{
\item{pt}{a  probability table for degrees}

\item{name}{a string}

\item{values}{a list of values}
}
\value{
the table with additional rows minus the original rows
}
\description{
Takes the probability table for matching passed as a parameter, 
and extends its rows with a novel variable
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
