% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{expand_with_variable_row.data.frame}
\alias{expand_with_variable_row.data.frame}
\title{Expands a data frame by adding combinations with a novel variable}
\usage{
expand_with_variable_row.data.frame(df, name, values)
}
\arguments{
\item{df}{a data frame}

\item{name}{a string}

\item{values}{a list of values}
}
\value{
the same dataframe with additional rows minus the original rows
}
\description{
Takes the data frame df passed as a parameter, and 
adds to rows the combinations of existing values and the novel
variable values. 
For instance if the rows where a=1 and a=2, and you add name=b
and values=list(3,4), the rows will be a=1,b=3 a=1,b=4 a=2,b=3
a2,b=4
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
