% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{expand_with_variables.dpp_degree_cpt}
\alias{expand_with_variables.dpp_degree_cpt}
\title{Expands a conditional probability table for degrees with additional variables}
\usage{
expand_with_variables.dpp_degree_cpt(cpt, names, dictionary)
}
\arguments{
\item{cpt}{a conditional probability table for degrees}

\item{names}{a list of strings}

\item{dictionary}{a dictionary in which the domains of each variable of names is described}
}
\value{
the CPT with additional columns minus the original columns
}
\description{
Takes the conditional probability table for degrees passed as a parameter, 
and extends it with various variables
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
