% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{expand_with_variables.dpp_matching_probas.B}
\alias{expand_with_variables.dpp_matching_probas.B}
\title{Expands the rows of matching probabilities with additional variables}
\usage{
expand_with_variables.dpp_matching_probas.B(pt, names, dictionary)
}
\arguments{
\item{pt}{a probability table for degrees}

\item{names}{a list of strings}

\item{dictionary}{a dictionary in which the domains of each variable of names is described}
}
\value{
the table with additional rows minus the original rows
}
\description{
Takes the probability table for matching passed as a parameter, 
and extends it with various variables
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
