% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{extract_attributes_values}
\alias{extract_attributes_values}
\title{Extracts attribute values from a string}
\usage{
extract_attributes_values(name)
}
\arguments{
\item{name}{the string to decode.}
}
\value{
a named list with attribute as key and value for each
}
\description{
From a string link "att1=v1,att2=v2",
returns a named list containing names "att1" and "att2"
and values "v1" and "v2". An error (\link{stop}) is raised
if the string is malformed.
}
\keyword{internal}
