% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{first.columns}
\alias{first.columns}
\title{Returns the first n columns of a dataframe and adds ...}
\usage{
first.columns(df, ncol = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{ncol}{the count of columns to keep}
}
\value{
the same dataframe with ncol columns and additional "..."
}
\description{
When we export a table, we sometimes want to truncate its columns 
so it can be entirely displayed and be readable. This function keeps the 
first ncol columns (such as can be done using df[1:ncol]), and adds 
as a last column "..." symbols to show the table was truncated. 
Notably used for exportation to LaTeX tables.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{latex}
