% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{first.lines}
\alias{first.lines}
\title{Returns the first n rows of a dataframe and adds ...}
\usage{
first.lines(df, nrow = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{nrow}{the count of rows to keep}
}
\value{
the same dataframe with nrow columns and additional "..."
}
\description{
When we export a table, we sometimes want to truncate its rows 
so it can be entirely displayed and be readable. This function keeps the 
first nrow rows (such as can be done using df[1:nrow,]), and adds 
as a last row "..." symbols to show the table was truncated. 
Notably used for exportation to LaTeX tables.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{latex}
