% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{fisher.pvalue.or.NA}
\alias{fisher.pvalue.or.NA}
\title{Quantifies the Fisher goodness of fit between two dataframes}
\usage{
fisher.pvalue.or.NA(x, y)
}
\arguments{
\item{x}{data ref}

\item{y}{data gen}
}
\value{
a scalar or NA if not possible
}
\description{
Computes fisher test using \code{\link{fisher.test}};
in case of error, returns NA
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
