% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{generate.hypothesis}
\alias{generate.hypothesis}
\title{generates the hypothesis to explore a given solution}
\usage{
generate.hypothesis(sol)
}
\arguments{
\item{sol}{a current solution
 @return a list of lists of strings}
}
\description{
Explores all the combinations of variables which are missing 
and might be set to user inputs.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
