% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{gof.freeman_tukey}
\alias{gof.freeman_tukey}
\title{Quantifies the goodness of fit between two dataframes}
\usage{
gof.freeman_tukey(df1, df2)
}
\arguments{
\item{df1}{data ref}

\item{df2}{data gen}
}
\value{
a scalar
}
\description{
Computes the squared of each, then the difference between them,
then elevated to power 2, then summed and mult 4.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
