% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heat_map_gradient}
\alias{heat_map_gradient}
\title{Creates a ggplot2 gradient adapted to the values for an heatmap}
\usage{
heat_map_gradient(d)
}
\arguments{
\item{d}{the data to plot}
}
\value{
a scale_fill_gradient2
}
\description{
It creates a hot/cold symetric gradient
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
