% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{inf_to_zeros}
\alias{inf_to_zeros}
\title{Replaces Infinite by zeros}
\usage{
inf_to_zeros(vv)
}
\arguments{
\item{vv}{a vector or list}
}
\value{
the same vector without Infinite
}
\description{
Replaces Infinite by zeros. Used when we divide 
by a value and know the result should be 0 
when the diviser is 0.
}
\keyword{internal}
