% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{ipf.2d}
\alias{ipf.2d}
\title{Applies Iterative Proportional Fitting on a 2d dataframe}
\usage{
ipf.2d(df, margins_cols, margins_rows, max.iterations = 1000,
  precision = 1e-10, verbose = FALSE)
}
\arguments{
\item{df}{the dataframe to reweight}

\item{margins_cols}{a vector containing the target marginals for cols}

\item{margins_rows}{a vector containing the target marginals for rows}

\item{max.iterations}{the maximum iterations before stopping (1000 by default)}

\item{precision}{the target precision, will stop once reached (defaults to 1e-6)}

\item{verbose}{displays more messages if TRUE (FALSE by default)}
}
\value{
the dataframe reweighted
}
\description{
Takes as inputs a dataframe representing a 2d matrix, 
and target marings for cols and rows. 
Returns a table reweighted so its sums fit margins.
}
\details{
Either falls back to a local implementation \code{\link{ipf.2d.stackoverflow}}, or to an
higher quality one from the mipfp package if it is installed.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
