% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{ipf.2d.stackoverflow}
\alias{ipf.2d.stackoverflow}
\title{A simple Iterated Proportional Fitting implementation}
\usage{
ipf.2d.stackoverflow(Margins_, seedAry, maxiter = 100, closure = 0.001,
  verbose = FALSE)
}
\description{
taken from https://stats.stackexchange.com/questions/59115/iterative-proportional-fitting-in-r
}
\seealso{
The entry point for IPF 2D: \code{\link{ipf.2d}}
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
