% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{list_variabes_modalities}
\alias{list_variabes_modalities}
\title{Extracts the list of modalities per variable}
\usage{
list_variabes_modalities(strs)
}
\arguments{
\item{strs}{a list of strings}
}
\value{
a named list with for each name (variable) contains the vector of modalities
}
\description{
From a list of strings, extracts the modalities possible 
for each variable.
}
\keyword{internal}
