% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{matching.fix}
\alias{matching.fix}
\title{Fix inputs parameters for DPP resolution}
\usage{
matching.fix(sample.A, sample.B, pdi, pdj, pij)
}
\arguments{
\item{sample.A}{the sample to use for population A created using \code{\link{create_sample}}}

\item{sample.B}{the sample to use for population B created using \code{\link{create_sample}}}

\item{pdi}{the distribution of degrees for population A created with \code{\link{create_degree_probabilities_table}}}

\item{pdj}{the distribution of degrees for population B created with \code{\link{create_degree_probabilities_table}}}

\item{pij}{the matching probabilities created with \code{\link{create_matching_probabilities_table}}}
}
\value{
a named list containing the fixed variables
}
\description{
This function can fix what is not constitent in the naming 
of attributes, such as attributes not provided in pdi, pdj or pij
but provided for another item.
}
\details{
This function is called by \code{\link{matching.prepare}} if the optional parameter \code{fix=TRUE} is used.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
