% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{matching.generate.add_degree}
\alias{matching.generate.add_degree}
\title{Adds a target degree column to a population 
based on the distribution of contigency for degrees passed as a parameter.}
\usage{
matching.generate.add_degree(samp, pop, n, ndx, verbose)
}
\arguments{
\item{samp}{the original sample (used for its dictionary)}

\item{pop}{the population to which the column will be added to}

\item{n}{the expected contigencies (used as a control)}

\item{ndx}{the contigencies for each degree}

\item{verbose}{if TRUE, detailed messages are printed}
}
\description{
Adds a target degree column to a population 
based on the distribution of contigency for degrees passed as a parameter.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
