% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{matching.generate}
\alias{matching.generate}
\title{Generates a synthetic population based on the resolution of the problem,
applied on two samples \code{sample.A} and \code{sample.B}. 
It first resizes the two populations so they reach the expected sizes, 
the adds degrees to them, then creates links based on degrees 
and characteristics.}
\usage{
matching.generate(case, sample.A, sample.B, verbose = FALSE,
  force = FALSE)
}
\arguments{
\item{case}{the case resolved by the \code{\link{matching.solve}} function.}

\item{sample.A}{the sample to use as population A}

\item{sample.B}{the sample to use as population B}

\item{verbose}{displays detailed messages if TRUE}

\item{force}{if TRUE, will try to generate even very big populations (and will likely fail)}
}
\value{
the generated population
}
\description{
Generates a synthetic population based on the resolution of the problem,
applied on two samples \code{sample.A} and \code{sample.B}. 
It first resizes the two populations so they reach the expected sizes, 
the adds degrees to them, then creates links based on degrees 
and characteristics.
}
\seealso{
\code{\link{matching.solve}} to prepare the case for this call
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
