% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{matching.prepare}
\alias{matching.prepare}
\title{Prepares a case for Direct Probabilistic Peering}
\usage{
matching.prepare(sample.A, sample.B, pdi, pdj, pij, fix = TRUE)
}
\arguments{
\item{sample.A}{the sample to use for population A created using \code{\link{create_sample}}}

\item{sample.B}{the sample to use for population B created using \code{\link{create_sample}}}

\item{pdi}{the distribution of degrees for population A created with \code{\link{create_degree_probabilities_table}}}

\item{pdj}{the distribution of degrees for population B created with \code{\link{create_degree_probabilities_table}}}

\item{pij}{the matching probabilities created with \code{\link{create_matching_probabilities_table}}}

\item{fix}{if \code{TRUE}, the function will fix mismatches (notably missing variables in pdi, pdj or pij) by itself}
}
\value{
a case ready to be solved with \code{\link{matching.solve}}
}
\description{
Prepares a case based on two samples and two sets of probabilities.
It takes two samples, probabilities distributions for degrees 
(count of links per entity) and matching probabilities. 
It ensures all the parameters are consistent. In case \code{fix=TRUE}, a prior temptative 
to fix minor inconsistencies will be done using \code{\link{matching.fix}}.
}
\examples{
data(dwellings_households)
prepared <- matching.prepare(dwellings_households$sample.A, dwellings_households$sample.B, 
                             dwellings_households$pdi, dwellings_households$pdj, 
                             dwellings_households$pij)

}
\seealso{
\code{\link{matching.solve}} to use the result of this function
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
