% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{mean.data.frame}
\alias{mean.data.frame}
\title{Computes the mean of all the values inside a dataframe}
\usage{
\method{mean}{data.frame}(x, ...)
}
\arguments{
\item{x}{the dataframe}

\item{...}{additional parameters are quietly ignored}
}
\value{
a double value
}
\description{
Sometimes the standard mean() function works on matrices,
but it does not on dataframes. In case the parameter is a dataframe,
the mean will be computed as the sum of all the values divided by the 
count of elements. Else another standard mean will be called.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
