% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{measure_contigencies}
\alias{measure_contigencies}
\title{Measures the contigencies from a sample}
\usage{
measure_contigencies(sample, strs)
}
\arguments{
\item{sample}{the sample to measure}

\item{strs}{the vector of strings containing the variable names and modalities to be measured}
}
\value{
a vector containing the contigencies
}
\description{
measures, for a given variable, the cumulated weights for 
each modality of this variable in the sample
}
\keyword{internal}
