% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{measure_pairing_from_df}
\alias{measure_pairing_from_df}
\title{Create pairing probabilities}
\usage{
measure_pairing_from_df(df, var.row, var.col, var.weight)
}
\arguments{
\item{df}{the dataframe}

\item{var.row}{the variables to use as rows}

\item{var.col}{the variables to use as cols}

\item{var.weight}{the name of the variable to use for weight}
}
\value{
a dataframe structured in two dimensions as joint probabilities
}
\description{
Takes a dataframe containing a sample, and extracts a joint probabilitie table
ready to be passed to \code{\link{create_matching_probabilities_table}}.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
