% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{merge_links}
\alias{merge_links}
\title{Merge entities and attributes}
\usage{
merge_links(pop)
}
\arguments{
\item{pop}{the generated population}
}
\value{
a dataframe from the join
}
\description{
Merges a generated population: returns the join of population A,
links and population B.
}
\details{
The underlying operation is done by \code{\link{merge}}.
}
\keyword{internal}
