% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{name.hypothesis}
\alias{name.hypothesis}
\title{A string representing an hypothesis}
\usage{
name.hypothesis(hypothesis)
}
\arguments{
\item{hypothesis}{a list of strings representing variables, as generated by \code{\link{generate.hypothesis}}
 @return a string}
}
\description{
For a given hypothesis l("hat.di","hat.fi"), 
returns a string like "hat.di=di, hat.fi=fi"
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
