% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{normalise}
\alias{normalise}
\title{Normalise an object}
\usage{
normalise(d, ...)
}
\arguments{
\item{d}{any data to normalise}

\item{...}{additional parameters are likely to be ignored}
}
\value{
the same object normalised
}
\description{
takes a vector, matrix or list and updates its content
by dividing it by its sum.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
