% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{paste.known}
\alias{paste.known}
\title{Concatenates the known values}
\usage{
paste.known(sol)
}
\arguments{
\item{sol}{the current solution (a list)}
}
\value{
a string
}
\description{
Only keeps in a solution (a list) the variables names which correspondond  to 
the chain of variables we need to solve
}
\keyword{internal}
