% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.dpp_resolved}
\alias{plot.dpp_resolved}
\alias{plot.dpp_result}
\title{Plots a visual synthese of the errors induced by the resolution process.}
\usage{
\method{plot}{dpp_resolved}(x, nameA = "A", nameB = "B",
  colorRef = "darkgray", colorSynthetic = "blue", ...)

\method{plot}{dpp_result}(x, nameA = "A", nameB = "B",
  colorRef = "darkgray", colorSynthetic = "blue", ...)
}
\arguments{
\item{x}{an object returned by the \code{\link{matching.solve}} or the \code{\link{matching.generate}} method}

\item{nameA}{a meaningfull label for the entity type of population A, such as "dwellings" (default to "A")}

\item{nameB}{a meaningfull label for the entity type of population B, such as "households" (default to "B")}

\item{colorRef}{the color to be used to plot values passed as parameters (defaults to "darkgray")}

\item{colorSynthetic}{the color to be used to plot values measured in the synthetic population (defaults to "blue")}

\item{...}{other parameters will be ignored quietly}
}
\value{
nothing
}
\description{
Plots all the values controlled by the algorithm, and compares their expected value (passed as data or parameter)
and their actual value (as measured in the resulting synthetic population).
The resulting graphs contain: 
\itemize{
 \item a graph showing the relaxation parameters passed to the solving function, as computed by \code{\link{plot_relaxation}}
 \item a graph showing the Normalized Root Mean Square Error (NRMSE) for the each control variable, as computed by \code{\link{plot_errors}}
 \item a graph showing the population sizes asked for and generated for populations A and B, as computed by \code{\link{plot_population_sizes}}
 \item a graph showing the difference between the input and observed pairing probabilities pij, as computed by \code{\link{plot_errors_pij}} 
 \item two graphs showing the initial and observed frequencies of control variables in both populations A and B, as computed by \code{\link{plot_frequencies_A}}
 \item two graphs showing the initial and observed average degrees in both populations A and B, as computed by \code{\link{plot_average_degree_A}}
 \item two graphs showing the difference between the expected and measured distribution of probability of degrees for both A and B, as computed by \code{\link{plot_errors_pdi}}
}
}
\details{
Note you can call the individual plotting functions for rendering for papers or zooming.
}
\examples{
data(dwellings_households)
prepared <- matching.prepare(
                     dwellings_households$sample.A, dwellings_households$sample.B, 
                     dwellings_households$pdi, dwellings_households$pdj, 
                     dwellings_households$pij)
solved <- matching.solve(
                     prepared, nA=50000, nB=40000, nu.A=1, phi.A=0, 
                     delta.A=0, gamma=0, delta.B=0, phi.B=0, nu.B=1, verbose=TRUE)
plot(solved, "dwellings", "households")
# we might as well drive the same on a generated case:
# sp <- matching.generate(
#                     solved, 
#                     sample.A=dwellings_households$sample.A, 
#                     sample.B=dwellings_households$sample.B, 
#                     verbose=TRUE)
# plot(sp, "dwellings", "households")

}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
