% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_average_degree_A}
\alias{plot_average_degree_A}
\alias{plot_average_degree_B}
\title{Plots average degrees, expected and generated, as bar charts}
\usage{
plot_average_degree_A(sp, nameA = "A", colorRef = "darkgray",
  colorSynthetic = "blue")

plot_average_degree_B(sp, nameB = "B", colorRef = "darkgray",
  colorSynthetic = "blue")
}
\arguments{
\item{sp}{a synthetic population, as produced by \code{\link{matching.generate}}.}

\item{nameA}{a meaningfull label for the entity type of population A, such as "dwellings" (default to "A")}

\item{colorRef}{the color to be used to plot values passed as parameters (defaults to "darkgray")}

\item{colorSynthetic}{the color to be used to plot values measured in the synthetic population (defaults to "blue")}

\item{nameB}{a meaningfull label for the entity type of population B, such as "households" (default to "B")}
}
\value{
a ggplot ready to display
}
\description{
Plots as a bar charts the avarage degree as they were expected and generated.
}
\seealso{
\code{\link{plot.dpp_result}} for the plot of all the results in one call
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
