% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_errors}
\alias{plot_errors}
\title{Plots the error measures}
\usage{
plot_errors(sp, colorSynthetic = "blue")
}
\arguments{
\item{sp}{a synthetic population, as produced by \code{\link{matching.generate}}.}

\item{colorSynthetic}{the color to be used to plot values measured in the synthetic population (defaults to "blue")}
}
\value{
a ggplot ready to display
}
\description{
Plots as a bar chart all the NRMSE (Normalized Rootsquared Mean Squarred Error) error measures
}
\seealso{
\code{\link{plot.dpp_result}} for the plot of all the results in one call
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
