% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_errors_pij}
\alias{plot_errors_pij}
\title{Plots expected and generated pairing probability}
\usage{
plot_errors_pij(sp, nameA = "A", nameB = "B")
}
\arguments{
\item{sp}{a synthetic population, as produced by \code{\link{matching.generate}} or \code{\link{matching.solve}}.}

\item{nameA}{a meaningfull label for the entity type of population A, such as "dwellings" (default to "A")}

\item{nameB}{a meaningfull label for the entity type of population B, such as "households" (default to "B")}
}
\value{
a ggplot ready to display
}
\description{
Plots as a heatmap the difference between expected pairing probabilities and 
generated ones. Locations in red have a lower probability than expected; in blue, it's the opposite.
}
\seealso{
\code{\link{plot.dpp_result}} for the plot of all the results in one call
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
