% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_population_sizes}
\alias{plot_population_sizes}
\title{Plots the expected and generated population sizes}
\usage{
plot_population_sizes(sp, nameA = "A", nameB = "B",
  colorRef = "darkgray", colorSynthetic = "blue")
}
\arguments{
\item{sp}{a synthetic population, as produced by \code{\link{matching.generate}}.}

\item{nameA}{a meaningfull label for the entity type of population A, such as "dwellings" (default to "A")}

\item{nameB}{a meaningfull label for the entity type of population B, such as "households" (default to "B")}

\item{colorRef}{the color to be used to plot values passed as parameters (defaults to "darkgray")}

\item{colorSynthetic}{the color to be used to plot values measured in the synthetic population (defaults to "blue")}
}
\value{
a ggplot ready to display
}
\description{
Plots as a bar chart the sizes expected and generated of the populations
}
\seealso{
\code{\link{plot.dpp_result}} for the plot of all the results in one call
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
