% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variable}
\alias{plot_variable}
\title{Plots the difference between proportions in the original sample and synthethic population}
\usage{
plot_variable(sample, generated, var.name, colorRef = "darkgray",
  colorSynthetic = "blue", as.factor = TRUE)
}
\arguments{
\item{sample}{a sample as produced by \code{\link{create_sample}}}

\item{generated}{a population as produced by \code{\link{matching.generate}}}

\item{var.name}{the name of the variable to measure and plot}

\item{colorRef}{the color to be used to plot values passed as parameters (defaults to "darkgray")}

\item{colorSynthetic}{the color to be used to plot values measured in the synthetic population (defaults to "blue")}

\item{as.factor}{if TRUE (default), plots the variable as a factor instead of numerical}
}
\value{
the plot
}
\description{
Measures the difference between the total weight of the various levels of a variable in the original sample,
and the actual count of each level in the synthetic population.
}
