% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{print.dpp_degree_cpt}
\alias{print.dpp_degree_cpt}
\title{Display a sample for Direct Probabilistic Peering}
\usage{
\method{print}{dpp_degree_cpt}(x, ...)
}
\arguments{
\item{x}{the degrees distribution of probabilities to display}

\item{...}{ignored}
}
\description{
Display a sample for Direct Probabilistic Peering
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
