% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{print.dpp_matching_probas}
\alias{print.dpp_matching_probas}
\title{Display matching probabilities}
\usage{
\method{print}{dpp_matching_probas}(x, ...)
}
\arguments{
\item{x}{the matching probabilities to print}

\item{...}{ignored}
}
\description{
Display matching probabilities
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
