% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{print.dpp_measure}
\alias{print.dpp_measure}
\title{Display the result of measures on a generated population}
\usage{
\method{print}{dpp_measure}(x, ...)
}
\arguments{
\item{x}{the measure to display}

\item{...}{ignored}
}
\description{
Display the result of measures on a generated population
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
