% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{print.dpp_population}
\alias{print.dpp_population}
\title{Display a generated population}
\usage{
\method{print}{dpp_population}(x, ...)
}
\arguments{
\item{x}{the population to print}

\item{...}{ignored}
}
\description{
Display a generated population
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
