% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{print.dpp_prepared}
\alias{print.dpp_prepared}
\title{Display a prepared case}
\usage{
\method{print}{dpp_prepared}(x, ...)
}
\arguments{
\item{x}{the case to print}

\item{...}{ignored}
}
\description{
Display a prepared case
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
