% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{print.dpp_resolved}
\alias{print.dpp_resolved}
\title{Display the result of a resolution}
\usage{
\method{print}{dpp_resolved}(x, ...)
}
\arguments{
\item{x}{the case to print}

\item{...}{ignored}
}
\description{
Display the result of a resolution
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
