% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{print.dpp_result}
\alias{print.dpp_result}
\title{Display a generated population}
\usage{
\method{print}{dpp_result}(x, ...)
}
\arguments{
\item{x}{the generation result to display}

\item{...}{ignored}
}
\description{
Display a generated population
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
