% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{propagate.direct}
\alias{propagate.direct}
\title{Completes a partial solution by inference}
\usage{
propagate.direct(sol, case, precision.pd = .Machine$double.eps,
  verbose = FALSE, indent = 1)
}
\arguments{
\item{sol}{the current solution (a named list)}

\item{case}{the case to solve}

\item{precision.pd}{the precision to be used for probabilistic distributions of degrees comparisons}

\item{verbose}{if TRUE, will display detailed information on the console}

\item{indent}{the level of indentation for verbose messages}
}
\value{
a solution with possibly more variables known
}
\description{
Takes an incomplete solution and enriches by applying the equations.
For instance if an equation links \code{hat.di * hat.ci = hat.ni}, 
then this equation will be applied if two of these variables are available.
}
\seealso{
this function relies on \code{\link{round_sum}} to round values 
         so their sums is preserved; 
         \code{\link{normalise}} to normalise variables;
         \code{\link{update_degree_distribution}} to compute the detail 
         of degree distributions from average degrees.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
