% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{quantify.errors}
\alias{quantify.errors}
\title{Quantifies the errors of a solution}
\usage{
quantify.errors(sol, case, nA, nB, compute.fisher = F,
  compute.chi2 = F)
}
\arguments{
\item{sol}{the solution to evaluate}

\item{case}{the reference case}

\item{nA}{the expected size of population A}

\item{nB}{the expected size of population B}

\item{compute.fisher}{if TRUE, the compute the Fisher p value (slow)}

\item{compute.chi2}{if TRUE, the compute the Chi2}
}
\value{
the solution with additional variables for errors
}
\description{
Computes the errors in a (possibly partial) solution 
for the variables which are defined and adds them 
to these solutions. For instance if \code{sol$hat.fi}
is available, \code{sol$mse.fi} will be computed.
}
\keyword{internal}
