% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{rectify.degree.counts}
\alias{rectify.degree.counts}
\title{Patches a degree distribution contingencies so its totals fit expectations}
\usage{
rectify.degree.counts(pdn, nn, cn, verbose = FALSE)
}
\arguments{
\item{pdn}{the contigencies of degrees}

\item{nn}{the expected total degree (ni) that should be reached}

\item{cn}{the expected total count (ci) that should be reached}

\item{verbose}{print detailed information during the process}
}
\description{
Adapts a distribution of degrees so that each column times n sums to expected nn
(basically solves rounding issues for this specific case)
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
