% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{relax.zeros}
\alias{relax.zeros}
\title{Relax the dataset by changing zeros with low probability}
\usage{
relax.zeros(x, by = .Machine$double.eps, ...)
}
\arguments{
\item{x}{the case to print}

\item{by}{the value to use to replace zeros (defaults to 1e-6)}

\item{...}{ignored}
}
\description{
Relax the dataset by changing zeros with low probability
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
