% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{remove_excess_probas_pd_lines}
\alias{remove_excess_probas_pd_lines}
\title{Removes useless modalities from a probability distribution}
\usage{
remove_excess_probas_pd_lines(pd, dictionary)
}
\arguments{
\item{pd}{a probability distribution}

\item{dictionary}{a dictionary}
}
\value{
the same probability distribution with maybe less elements
}
\description{
Often, the user provides probabilities to deal with all the possible modalities
in its dataset; but sometimes the sample just does not contains some of the 
modalities which might exist. This function takes as a reference the dictionary 
assuming it contains only what is required. In removes all the rows related
to modalities not really found in the probability table.
}
\details{
If data is removed, a warning will be emitted.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
