% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{reorder_column_modalities}
\alias{reorder_column_modalities}
\title{Reorders the modalities inside columns of a dataset}
\usage{
reorder_column_modalities(dataset, expected_order)
}
\arguments{
\item{dataset}{a dataset containing a dataframe}

\item{expected_order}{a list of strings corresponding to the modalities names}
}
\value{
the same dataset with different names for the same data
}
\description{
The user, or the automatic fixing process, might lead to situations where 
a pdi has for variable "b=1,a=2" whilst the order is "a=2,b=1" in other datasets.
As all the solving process relies on the hypothesis the order is OK, 
checking of such a dataset would fail.
}
\details{
This function helps fixing these situations by just changing the order of
modalities inside each column name
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
