% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorders the columns of a dataset so they are in the expected order}
\usage{
reorder_columns(df, expected_order)
}
\arguments{
\item{df}{a data frame}

\item{expected_order}{a list of strings corresponding to the same col names with a different order}
}
\value{
the data frame in a different order
}
\description{
The user, or the automatic fixing process, might lead to situations where 
a pdi has for variable list(a=1, a=2) whilst pij has list(a=2,a=1). 
As all the solving process relies on the hypothesis the order is OK, 
checking of such a dataset would fail.
}
\details{
This function helps fixing these situations by just changing the order of
the dataset's columns
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
