% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve.missing.chain}
\alias{resolve.missing.chain}
\title{Solves a chain of missing variables}
\usage{
resolve.missing.chain(sol, chain, case, nA, nB, nu.A, phi.A, delta.A, nu.B,
  phi.B, delta.B, gamma, verbose = FALSE, tolerance.pd = 1.5e-08,
  tolerance.degree.max = 1.5e-08, tolerance.pij = 1e-06)
}
\arguments{
\item{sol}{the current solution (a named list)}

\item{chain}{the chain to be solved (a list of strings containing variable names)}

\item{case}{the case to solve}

\item{nA}{the target population size for A}

\item{nB}{the target population size for B}

\item{nu.A}{control for nA: 0 means "respect nA", non-null "adapt it to solve the case"}

\item{phi.A}{control for frequencies: 0 means "respect the original frequencies as detected in the sample", non-null "adapt it to solve the case"}

\item{delta.A}{control for degree A: 0 means "respect the input parameters pdi", non-null "adapt them to solve the case"}

\item{nu.B}{control for nB: 0 means "respect nB", non-null "adapt it to solve the case"}

\item{phi.B}{control for frequencies: 0 means "respect the original frequencies as detected in the sample", non-null "adapt it to solve the case"}

\item{delta.B}{control for degree B: 0 means "respect the input parameters pdj", non-null "adapt them to solve the case"}

\item{gamma}{control for pij: 0 means "respect the matching probabilities pij", non-null "adapt them to solve the case"}

\item{verbose}{if TRUE, will display detailed information on the console}

\item{tolerance.pd}{the tolerance for probability distributions}

\item{tolerance.degree.max}{the tolerance for min/max average degrees}

\item{tolerance.pij}{the tolerance for pij probabilities}
}
\value{
a list of vectors (the chains) of strings
}
\description{
For a given incomplete solution, a set of dependant variables to solve,
 a generation case, and targets defined by the user, tries to identify 
the best solution (if any).
}
\details{
It works by (i) trying to define hypothesis such as "let's try to respect 
the ideal value for this variable". It then (ii) relies on inference to 
determine the consequences of this hypothesis, and checks (iii) whether
this solution is consistent.

At the end of this process, we might have no solution, one unique solution or 
several ones. If several solutions are available, the best one is taken, with best
being defined as the solution which minimizes the cumulated errors weighted by the 
user weights.
}
\seealso{
\code{\link{propagate.direct}} for the inference of the consequences of the hypothesis,
         \code{\link{detect.problems}} to ensure potential solutions are consistent
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
