% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\title{Resolves a partial solution by inference and hypothesis}
\usage{
resolve(sol, case, nA, nB, nu.A, phi.A, delta.A, gamma, nu.B, phi.B,
  delta.B, verbose = FALSE, tolerance.pd = 1.5e-08,
  tolerance.degree.max = 1.5e-08, tolerance.pij = 1e-06)
}
\arguments{
\item{sol}{the current solution (a named list)}

\item{case}{the case to solve}

\item{nA}{the target population size for A}

\item{nB}{the target population size for B}

\item{nu.A}{control for nA: 0 means "respect nA", non-null "adapt it to solve the case"}

\item{phi.A}{control for frequencies: 0 means "respect the original frequencies as detected in the sample", non-null "adapt it to solve the case"}

\item{delta.A}{control for degree A: 0 means "respect the input parameters pdi", non-null "adapt them to solve the case"}

\item{gamma}{control for pij: 0 means "respect the matching probabilities pij", non-null "adapt them to solve the case"}

\item{nu.B}{control for nB: 0 means "respect nB", non-null "adapt it to solve the case"}

\item{phi.B}{control for frequencies: 0 means "respect the original frequencies as detected in the sample", non-null "adapt it to solve the case"}

\item{delta.B}{control for degree B: 0 means "respect the input parameters pdj", non-null "adapt them to solve the case"}

\item{verbose}{if TRUE, will display detailed information on the console}

\item{tolerance.pd}{the tolerance for probability distributions}

\item{tolerance.degree.max}{the tolerance for min/max average degrees}

\item{tolerance.pij}{the tolerance for pij probabilities}
}
\value{
a list of vectors (the chains) of strings
}
\description{
Resolves a given case starting with the base solution
according to user parameters. 
It first uses \code{\link{propagate.direct}} to infer 
results from available elements; then it detects the 
chains of variable which are not constrained enough to be solved this way 
using \code{\link{detect.missing.chains}}. Then it solves each chain 
iteratively until the entire problem is solved.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
