% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{round_sum}
\alias{round_sum}
\title{Rounds an object by ensuring its totals remains equal before and after rounding}
\usage{
round_sum(x, ...)
}
\arguments{
\item{x}{the object to round}

\item{...}{ignored}
}
\value{
the rounded version of the object
}
\description{
Freely inspired by \url{https://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}
}
