% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{set.rownames.degrees}
\alias{set.rownames.degrees}
\title{Sets row names for degrees distribution}
\usage{
set.rownames.degrees(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
the dataframe with fixed names
}
\description{
Changes the row names so their numbering
starts at 0 instead of 1. It's fitting the
actual semantics of this type of table.
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
