% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{sum.list}
\alias{sum.list}
\title{Sums applied on a list}
\usage{
\method{sum}{list}(l)
}
\arguments{
\item{l}{the list to sum}
}
\value{
the same object summed with operator "+"
}
\description{
sums the content of the list (required for recent R versions)
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
