% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{update_degree_distribution.col}
\alias{update_degree_distribution.col}
\title{Update a given distribution column of degrees probabilities 
so its matching a target average degree for this column.}
\usage{
update_degree_distribution.col(pdx, t.target, verbose = FALSE,
  precision = 1e-10)
}
\arguments{
\item{pdx}{the distribution of degrees}

\item{t.target}{the vector of the expected average degrees}

\item{verbose}{more text if TRUE (FALSE by default)}

\item{precision}{for comparisons}
}
\description{
It works by finding a pivot and then increasing values on one 
side and removing probabilities on the other side.
}
\author{
samuel.thiriot@res-ear.ch
}
\keyword{internal}
