% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{update_degree_distribution}
\alias{update_degree_distribution}
\title{Update a given distribution of degrees probabilities so its matching target degrees}
\usage{
update_degree_distribution(pdx, dx, precision = .Machine$double.eps,
  verbose = F)
}
\arguments{
\item{pdx}{the distribution of degrees}

\item{dx}{the vector of the expected average degrees}

\item{precision}{the precision for comparison with 0; change only if suggested and relevant}

\item{verbose}{more messages if TRUE}
}
\description{
This process is done column by column by calling \link{update_degree_distribution.col}.
'
}
\author{
Samuel Thiriot <samuel.thiriot@res-ear.ch>
}
\keyword{internal}
